requirements.txt
aniso8601==7.0.0
astroid==2.2.5
attrs==19.1.0
bleach==3.1.0
bokeh==1.3.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.11
django-allauth==0.38.0
django-contrib-comments==1.9.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-generic-aggregation==0.4.0
django-haystack==2.8.1
django-ratelimit-backend==2.0
django-recaptcha==2.0.2
django-taggit==0.23.0
django-tinymce4-lite==1.7.5
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
flask-restplus==0.13.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jsmin==2.2.2
jsonschema==3.0.2
lazy-object-proxy==1.4.2
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.2
oauthlib==3.1.0
packaging==19.1
pandas==0.25.1
Pillow==5.4.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.4.2
pyrsistent==0.15.4
python-akismet==0.4.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.21.0
requests-oauthlib==1.2.0
six==1.11.0
sqlparse==0.3.0
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.15.6
whitenoise==4.1.2
Whoosh==2.7.4
wrapt==1.11.2
